+++ src/apps/ScopeChangeRequest/ScopeChangeRequestApp.tsx
@@ -37,28 +37,32 @@ export function setup(appApi: ClientApi): void {
         .registerStatusItems(statusBarConfig)
         .registerFilterOptions(filterConfig)
         .registerPrefetchQueries(prefetchQueriesOptions)
-        // .registerHelpPage({ Component: ScopeChangeRequestHelpPage })
         .registerSearchOptions([
             { name: 'Id', valueFormatter: ({ sequenceNumber }) => sequenceNumber.toString() },
             { name: 'Title', valueFormatter: ({ title }) => title },
+            { name: 'Description', valueFormatter: ({ description }) => description },
+            {
+                name: 'Warranty case',
+                valueFormatter: (s) => (s.potentialWarrantyCase ? 'Warranty' : ''),
+            },
         ])
         .registerPrefetchQueries(prefetchQueriesOptions)
         .registerPowerBIOptions(
             !appApi.isProduction
                 ? {
-                    pages: [
-                        {
-                            pageId: 'ReportSection4a8a96cf641d50747a13',
-                            pageTitle: 'Overview',
-                            default: true,
-                        },
-                        {
-                            pageId: 'ReportSection40a8a70e6f82243888ca',
-                            pageTitle: 'History',
-                        },
-                    ],
-                    reportURI: 'pp-scope-change-analytics',
-                }
+                      pages: [
+                          {
+                              pageId: 'ReportSection4a8a96cf641d50747a13',
+                              pageTitle: 'Overview',
+                              default: true,
+                          },
+                          {
+                              pageId: 'ReportSection40a8a70e6f82243888ca',
+                              pageTitle: 'History',
+                          },
+                      ],
+                      reportURI: 'pp-scope-change-analytics',
+                  }
                 : (undefined as unknown as PowerBiOptions)
         );
 }
